# lib/api-zod/src/generated/types/createMyOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tata Pizza API - Sistema de automatización de pedidos con portal cliente
 * OpenAPI spec version: 0.2.0
 */
import type { CreateOrderItemBody } from "./createOrderItemBody";

export interface CreateMyOrderBody {
  /** @nullable */
  deliveryAddress?: string | null;
  deliveryType: string;
  paymentMethod: string;
  /** @nullable */
  notes?: string | null;
  items: CreateOrderItemBody[];
}