# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tata Pizza API - Sistema de automatización de pedidos con portal cliente
 * OpenAPI spec version: 0.2.0
 */
import type { CreateOrderItemBody } from "./createOrderItemBody";

export interface CreateOrderBody {
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  deliveryAddress?: string | null;
  deliveryType?: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  whatsappMessageId?: string | null;
  items: CreateOrderItemBody[];
}