# lib/api-zod/src/generated/types/createRatingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tata Pizza API - Sistema de automatización de pedidos con portal cliente
 * OpenAPI spec version: 0.2.0
 */

export interface CreateRatingBody {
  productRating: number;
  deliveryRating: number;
  serviceRating: number;
  /** @nullable */
  comment?: string | null;
}