# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tata Pizza API - Sistema de automatización de pedidos con portal cliente
 * OpenAPI spec version: 0.2.0
 */
import type { Rating } from "./rating";

export interface Order {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  clerkUserId?: string | null;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  deliveryAddress?: string | null;
  deliveryType: string;
  status: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  paymentStatus?: string | null;
  /** @nullable */
  paymentProofUrl?: string | null;
  total: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  whatsappMessageId?: string | null;
  rating?: Rating | null;
  createdAt: Date;
  updatedAt: Date;
}