# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tata Pizza API - Sistema de automatización de pedidos con portal cliente
 * OpenAPI spec version: 0.2.0
 */

export interface OrderItem {
  id: number;
  orderId: number;
  /** @nullable */
  menuItemId?: number | null;
  name: string;
  /** @nullable */
  size?: string | null;
  quantity: number;
  unitPrice: number;
  subtotal: number;
  createdAt: Date;
}