# lib/api-zod/src/generated/types/updateAddressBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tata Pizza API - Sistema de automatización de pedidos con portal cliente
 * OpenAPI spec version: 0.2.0
 */

export interface UpdateAddressBody {
  label?: string;
  address?: string;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  reference?: string | null;
  /** @nullable */
  instructions?: string | null;
  isDefault?: boolean;
}