# lib/api-zod/src/generated/types/updateOrderPaymentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tata Pizza API - Sistema de automatización de pedidos con portal cliente
 * OpenAPI spec version: 0.2.0
 */

export interface UpdateOrderPaymentBody {
  paymentStatus: string;
  /** @nullable */
  paymentProofUrl?: string | null;
}